Refactor to decrease layers
